Uploaded edits to code file
Now has the ability to sort by basic climate variables as well. It's definitely clunky, but I hope to use this with some sort of GUI, so that, for example, a user could check boxes for the climate variables (as well as plant types, for that matter) and not have to type specific words to get it to work.